Extract pseudopackage for dialogs
