+++ packages/dialog/InputDialog.js
