+++ packages/dialog/index.js
